+++ src/app/dashboard/shared/pipe/option-is-selected.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { OptionIsSelectedPipe } from './option-is-selected.pipe';
+
+describe('OptionIsSelectedPipe', () => {
+  it('create an instance', () => {
+    const pipe = new OptionIsSelectedPipe();
+    expect(pipe).toBeTruthy();
+  });
+});